% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncrementVersion.R
\name{versionIncrement}
\alias{versionIncrement}
\title{increment package version number for a given patch or update.}
\usage{
versionIncrement(wd = getwd(), major = FALSE, mid = FALSE,
  minor = TRUE, maxdigits = c(2, 2, 3), README = TRUE)
}
\arguments{
\item{wd}{file path to package folder, ending in package name.}

\item{major}{logical. Shall we increment the major version number? Default FALSE}

\item{mid}{logical. Shall we increment the middle version number? Default FALSE}

\item{minor}{logical. Shall we increment the minor version number? Default TRUE}

\item{maxdigits}{integer vector of length 3, default \code{c(2,2,3)}.}

\item{README}{logical. Update the development version badge in the README, see details. Default TRUE.}
}
\value{
nothing is returned. Package called for the side effect of modifying the date and version elements of the \code{DESCRIPTION} file and possibly the \code{README.md} file.
}
\description{
Increment the version number, and update the DESCRIPTION and README.md files accordingly.
}
